moved tomb interfaces and enum to /types
