+++ src/components/shared/ResurrectionForm/ResurrectionTimeForm/CustomTimeSelect.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import DatePicker from 'react-datepicker'
 import { convertToUTCTime } from '../../components.utils'
-import { RewrapFormState } from '../../../SarcophagusTomb/tomb.interfaces'
 import { utcToDateObject } from '../../../SarcophagusTomb/tomb.utils'
 import { SarcophagusCreateValues } from '../../../../types/sarcophagusCreate'
+import { RewrapFormState } from '../../../../types/sarcophagusTomb'
 
 interface CustomTimeSelectProps {
   values: RewrapFormState | SarcophagusCreateValues;
